chore(ci): get CI working again (#759)
* update actions/upload-artifact to v4

* update actions/checkout to v4

* update actions/setup-node to v4

* pin versions for windows image

* update macos runner image

* disable code signing for windows

ideally this is temporary, as our current certificate has expired

* disable code signing on macOS

ideally temporary, as our certificate has expired

* properly disable code signing on macOS

* update electron-builder to v23.6.0

Prior to v23.0.2, error would occur on macOS due to python 2.7 being
removed by OS update (e.g.
https://github.com/electron-userland/electron-builder/issues/6726).

The alternative approach is updating the workflow to install and use
Python 2. Probably safer and less intrusive to do that, so may consider
switching to that approach if this upgrade causes other issues.

* pin ubuntu image version

* re-enable code signing

Certificate is updated in repo

* re-enable notarizing

* fix notarization

* add missing APPLE_TEAM_ID env

* code sign for pull request

---------

Co-authored-by: Gregor MacLennan <gmaclennan@awana.digital>